ZStream drop and take improvements (#2524)
* Make ZStream#drop take a Long argument

* Check that ZStream#drop is safe to pull again

* Make ZStream#take take a Long attribute

* Make ZStream#take safe to pull again

* Fix rebase issues

* Fix Scala 2.11 type inference issue

+++ streams-tests/jvm/src/test/scala/zio/stream/StreamSpec.scala
@@ -1621,7 +1621,7 @@ object StreamSpec extends ZIOBaseSpec {
     suite("Stream.take")(
       testM("take")(checkM(streamOfBytes, Gen.anyInt) { (s: Stream[String, Byte], n: Int) =>
         for {
-          takeStreamResult <- s.take(n).runCollect.run
+          takeStreamResult <- s.take(n.toLong).runCollect.run
           takeListResult   <- s.runCollect.map(_.take(n)).run
         } yield assert(takeListResult.succeeded)(isTrue) implies assert(takeStreamResult)(equalTo(takeListResult))
       }),
